Settings load and write to SPIFFS

+++ lib/Settings/Settings.h
@@ -26,19 +26,19 @@ struct Settings
 public:
   char host[40];
   char my_hostname[40];
-  int8_t restart_flag;
-  int8_t timezone;
-  uint16_t altitude;
+  int restart_flag;
+  int timezone;
+  int altitude;
   char mqtt_host[33];
-  uint16_t mqtt_port;
+  int mqtt_port;
   char mqtt_client[33];
   char mqtt_user[33];
   char mqtt_pwd[33];
   char mqtt_topic[33];
   float temp_offset;
-  uint16_t dry;
-  uint16_t wet;
-  int8_t loglevel;
+  int dry;
+  int wet;
+  int loglevel;
 };
 
 #endif
